Made changes based on code review
