+++ i18n/locales/en/common.json
@@ -1,5 +1,5 @@
 {
-  "\r\n                      You can choose a high, medium, or low transaction priority, each requiring a\r\n                      corresponding transaction fee. Or you can pay any desired fee of no less than\r\n                      {{minFee}} {{token}}. If you don't know what fee to pay, choose\r\n                      one of the provided transaction priorities.\r\n                    ": "\r\n                      You can choose a high, medium, or low transaction priority, each requiring a\r\n                      corresponding transaction fee. Or you can pay any desired fee of no less than\r\n                      {{minFee}} {{token}}. If you don't know what fee to pay, choose\r\n                      one of the provided transaction priorities.\r\n                    ",
+  "\n                      You can choose a high, medium, or low transaction priority, each requiring a\n                      corresponding transaction fee. Or you can pay any desired fee of no less than\n                      {{minFee}} {{token}}. If you don't know what fee to pay, choose\n                      one of the provided transaction priorities.\n                    ": "\n                      You can choose a high, medium, or low transaction priority, each requiring a\n                      corresponding transaction fee. Or you can pay any desired fee of no less than\n                      {{minFee}} {{token}}. If you don't know what fee to pay, choose\n                      one of the provided transaction priorities.\n                    ",
   " - Confirm transaction on your {{deviceModel}}": " - Confirm transaction on your {{deviceModel}}",
   " is out. ": " is out. ",
   " to deposit LSK.": " to deposit LSK.",
@@ -245,7 +245,7 @@
   "Lisk Terms of Use": "Lisk Terms of Use",
   "Lisk Website": "Lisk Website",
   "Lisk accounts on {{WalletModel}}": "Lisk accounts on {{WalletModel}}",
-  "Lisk counts your message in bytes, so keep in mind\r\n                that the length of your message may vary in different languages.\r\n                Different characters may consume a varying amount of bytes.": "Lisk counts your message in bytes, so keep in mind\r\n                that the length of your message may vary in different languages.\r\n                Different characters may consume a varying amount of bytes.",
+  "Lisk counts your message in bytes, so keep in mind\n                that the length of your message may vary in different languages.\n                Different characters may consume a varying amount of bytes.": "Lisk counts your message in bytes, so keep in mind\n                that the length of your message may vary in different languages.\n                Different characters may consume a varying amount of bytes.",
   "Lisk currently supports Ledger Nano S, Ledger Nano X, Trezor One and Trezor Model T wallets": "Lisk currently supports Ledger Nano S, Ledger Nano X, Trezor One and Trezor Model T wallets",
   "Lisk currently supports Ledger Nano S, Ledger Nano X, Trezor One and Trezor T wallets.": "Lisk currently supports Ledger Nano S, Ledger Nano X, Trezor One and Trezor T wallets.",
   "Lisk has now been enhanced even further": "Lisk has now been enhanced even further",
