+++ src/store/actions/transactions.test.js
@@ -11,7 +11,8 @@ import {
   transactionBroadcasted,
   transactionCreated,
 } from './transactions';
-import { sampleTransaction, transformedAccountTransaction, newTransaction } from '../../../test/constants/transactions';
+import { sampleTransaction } from '../../../test/constants/transactions';
+import { getState, transformedAccountTransaction, newTransaction } from '../../../test/fixtures/transactions';
 
 jest.mock('@api/transaction');
 jest.mock('@api/delegate');
@@ -24,47 +25,10 @@ describe('actions: transactions', () => {
   });
 
   const dispatch = jest.fn();
-  const getState = () => ({
-    account: {
-      passphrase: 'test',
-      info: {
-        LSK: {
-          summary: { address: 'lsks6uckwnap7s72ov3edddwgxab5e89t6uy8gjt6' },
-        },
-        BTC: {
-          summary: { address: '16Qp9op3fTESTBTCACCOUNTv52ghRzYreUuQ' },
-        },
-      },
-      loginType: 0,
-    },
-    network: {
-      status: { online: true },
-      name: 'Mainnet',
-      networks: {
-        LSK: {
-          serviceUrl: 'http://localhost:4000',
-          nethash: '198f2b61a8eb95fbeed58b8216780b68f697f26b849acf00c8c93bb9b24f783d',
-        },
-      },
-    },
-    transactions: {
-      filters: {},
-    },
-    settings: {
-      token: {
-        active: 'LSK',
-      },
-    },
-    blocks: {
-      latestBlocks: [{
-        timestamp: 123123123,
-      }],
-    },
-  });
 
   describe('getTransactions', () => {
     const data = {
-      address: '15626650747375562521L',
+      address: 'lsks6uckwnap7s72ov3edddwgxab5e89t6uy8gjt6',
       limit: 20,
       offset: 0,
       filters: {},
@@ -75,6 +39,7 @@ describe('actions: transactions', () => {
     });
 
     it('should dispatch getTransactionsSuccess action if resolved', async () => {
+      // Arrange
       transactionsApi.getTransactions.mockResolvedValue({ data: [], meta: { total: 0 } });
       const expectedAction = {
         count: 0,
@@ -84,16 +49,28 @@ describe('actions: transactions', () => {
         offset: 0,
       };
 
+      // Act
       await transactionsRetrieved(data)(dispatch, getState);
-      expect(dispatch).toHaveBeenLastCalledWith({
+
+      // Assert
+      expect(dispatch).toHaveBeenNthCalledWith(1, {
+        type: actionTypes.loadingStarted,
+        data: actionTypes.transactionsRetrieved,
+      });
+      expect(dispatch).toHaveBeenNthCalledWith(2, {
         type: actionTypes.transactionsRetrieved,
         data: expectedAction,
+      });
+      expect(dispatch).toHaveBeenNthCalledWith(3, {
+        type: actionTypes.loadingFinished,
+        data: actionTypes.transactionsRetrieved,
       });
     });
 
     it('should dispatch getTransactionsSuccess action if resolved and default argument values are used', async () => {
+      // Arrange
       const transactionData = {
-        address: '15626650747375562521L',
+        address: 'lsks6uckwnap7s72ov3edddwgxab5e89t6uy8gjt6',
       };
       transactionsApi.getTransactions.mockResolvedValue({
         data: transactionData,
@@ -107,22 +84,36 @@ describe('actions: transactions', () => {
         offset: 0,
       };
 
+      // Act
       await transactionsRetrieved(transactionData)(dispatch, getState);
-      expect(dispatch).toHaveBeenLastCalledWith({
+
+      // Assert
+      expect(dispatch).toHaveBeenNthCalledWith(1, {
+        type: actionTypes.loadingStarted,
+        data: actionTypes.transactionsRetrieved,
+      });
+      expect(dispatch).toHaveBeenNthCalledWith(2, {
         type: actionTypes.transactionsRetrieved,
         data: expectedAction,
+      });
+      expect(dispatch).toHaveBeenNthCalledWith(3, {
+        type: actionTypes.loadingFinished,
+        data: actionTypes.transactionsRetrieved,
       });
     });
 
     it('should dispatch transactionsLoadFailed action if rejected', async () => {
+      // Arrange
       const transactionError = new Error('Transaction retrieve error');
       transactionsApi.getTransactions.mockRejectedValue(transactionError);
 
-      expect(transactionsApi.getTransactions).rejects.toThrow('Transaction retrieve error');
-
+      // Act
       await transactionsRetrieved(data)(dispatch, getState);
+
+      // Assert
+      expect(transactionsApi.getTransactions).rejects.toThrow('Transaction retrieve error');
       expect(transactionsApi.getTransactions).rejects.toThrow(transactionError);
-      expect(dispatch).toHaveBeenCalledTimes(2);
+      expect(dispatch).toHaveBeenCalledTimes(3);
       expect(dispatch).toHaveBeenNthCalledWith(1, {
         type: actionTypes.loadingStarted,
         data: actionTypes.transactionsRetrieved,
@@ -130,6 +121,10 @@ describe('actions: transactions', () => {
       expect(dispatch).toHaveBeenNthCalledWith(2, {
         type: actionTypes.transactionLoadFailed,
         data: { error: transactionError },
+      });
+      expect(dispatch).toHaveBeenNthCalledWith(3, {
+        type: actionTypes.loadingFinished,
+        data: actionTypes.transactionsRetrieved,
       });
     });
   });
@@ -143,6 +138,7 @@ describe('actions: transactions', () => {
 
   describe('pendingTransactionAdded', () => {
     it('should create an action to add pending transaction', () => {
+      // Arrange
       const data = {
         moduleAssetId: '1938573839:g45krEIjwK',
         id: '4emF3me9YJSbcIuOp',
@@ -159,6 +155,8 @@ describe('actions: transactions', () => {
         type: actionTypes.pendingTransactionAdded,
         data,
       };
+
+      // Assert
       expect(pendingTransactionAdded(data)).toEqual(expectedAction);
     });
   });
@@ -172,31 +170,42 @@ describe('actions: transactions', () => {
 
   describe('transactionCreated', () => {
     it('should dispatch transactionCreatedSuccess action if there are no errors', async () => {
+      // Arrange
       transactionsApi.create.mockResolvedValue({ tx: newTransaction });
       const expectedAction = {
         type: actionTypes.transactionCreatedSuccess,
         data: { tx: newTransaction },
       };
+
+      // Act
       await transactionCreated(newTransaction)(dispatch, getState);
+
+      // Assert
       expect(transactionsApi.create).toHaveBeenCalled();
       expect(dispatch).toHaveBeenCalledWith(expectedAction);
     });
 
     it('should dispatch transactionSignError action if there are errors', async () => {
+      // Arrange
       const transactionError = new Error('Transaction create error');
       loginTypes.passphrase.code = 1;
       hwManagerApi.signSendTransaction.mockRejectedValue(transactionError);
-      expect(hwManagerApi.signSendTransaction).rejects.toThrow(transactionError);
       const expectedAction = {
         type: actionTypes.transactionSignError,
         data: transactionError,
       };
+
+      // Act
       await transactionCreated(newTransaction)(dispatch, getState);
+
+      // Assert
+      expect(hwManagerApi.signSendTransaction).rejects.toThrow(transactionError);
       expect(hwManagerApi.signSendTransaction).toHaveBeenCalled();
       expect(dispatch).toHaveBeenCalledWith(expectedAction);
     });
 
     it('should dispatch transactionSignError action if there are no transaction signatures', async () => {
+      // Arrange
       loginTypes.passphrase.code = 1;
       const noSignatureTransaction = {
         pending: [],
@@ -209,7 +218,11 @@ describe('actions: transactions', () => {
         type: actionTypes.transactionSignError,
         data: null,
       };
+
+      // Act
       await transactionCreated(newTransaction)(dispatch, getState);
+
+      // Assert
       expect(hwManagerApi.signSendTransaction).toHaveBeenCalled();
       expect(dispatch).toHaveBeenCalledWith(expectedAction);
     });
@@ -217,6 +230,7 @@ describe('actions: transactions', () => {
 
   describe('transactionDoubleSigned', () => {
     it('should create an action to reset transaction result', () => {
+      // Arrange
       const data = {
         id: 'a66c3cb626dbb631e53a5bd617cc8c13f06186c91f42e87008331e6dc4dc3cba',
         signatures: ['ofh49fhfnYHK8499dvvn', '0egeJT7dgdnek4n5k6994k'],
@@ -226,12 +240,15 @@ describe('actions: transactions', () => {
         type: actionTypes.transactionDoubleSigned,
         data,
       };
+
+      // Assert
       expect(transactionDoubleSigned(data)).toEqual(expectedAction);
     });
   });
 
   describe('transactionBroadcasted', () => {
     it('should dispatch broadcastedTransactionSuccess action if there are no errors', async () => {
+      // Arrange
       transactionsApi.broadcast.mockResolvedValue({ data: sampleTransaction });
       transactionsModifyApi.transformTransaction.mockImplementation(
         () => transformedAccountTransaction,
@@ -248,17 +265,19 @@ describe('actions: transactions', () => {
         data: new Date(),
       };
 
+      // Act
       await transactionBroadcasted(sampleTransaction)(dispatch, getState);
+
+      // Assert
       expect(transactionsApi.broadcast).toHaveBeenCalled();
       expect(dispatch).toHaveBeenCalledWith(expectedAction);
       expect(dispatch).toHaveBeenLastCalledWith(lastExpectedAction);
     });
 
     it('should dispatch broadcastedTransactionError action if there\'s an error', async () => {
+      // Arrange
       const transactionBroadcastError = new Error('Transaction broadcast error');
       transactionsApi.broadcast.mockRejectedValue(transactionBroadcastError);
-
-      expect(transactionsApi.broadcast).rejects.toThrow(transactionBroadcastError);
       const expectedAction = {
         type: actionTypes.broadcastedTransactionError,
         data: {
@@ -266,20 +285,29 @@ describe('actions: transactions', () => {
           transaction: sampleTransaction,
         },
       };
-      expect(transactionsApi.broadcast).toHaveBeenCalled();
+
+      // Act
       await transactionBroadcasted(sampleTransaction)(dispatch, getState);
+
+      // Assert
+      expect(transactionsApi.broadcast).rejects.toThrow(transactionBroadcastError);
+      expect(transactionsApi.broadcast).toHaveBeenCalled();
       expect(dispatch).toHaveBeenCalledWith(expectedAction);
     });
 
     it('should dispatch pendingTransactionAdded action if transformedTransaction sender address is the same as LSK address', async () => {
+      // Arrange
       transactionsApi.broadcast.mockResolvedValue({ data: sampleTransaction });
       transactionsModifyApi.transformTransaction.mockReturnValue(transformedAccountTransaction);
-
       const expectedAction = {
         type: actionTypes.pendingTransactionAdded,
         data: { ...transformedAccountTransaction, isPending: true },
       };
+
+      // Act
       await transactionBroadcasted(sampleTransaction)(dispatch, getState);
+
+      // Assert
       expect(transactionsModifyApi.transformTransaction).toHaveBeenCalled();
       expect(dispatch).toHaveBeenCalledWith(expectedAction);
     });
